LayoutTests/java/lc3/ConvertJSObject/ToInt-001-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion Preferred argument conversion:  JavaScript Object to int

FAIL TEST_CLASS.ambiguous( new String() ) +'' should be INT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new Boolean() ) +'' should be INT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new Number() ) +'' should be INT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new Date(0) ) +'' should be INT. Was BYTE.
PASS successfullyParsed is true

TEST COMPLETE
 
